requirements_unix.txt
# VoiceFlow Unix Dependencies (Linux/macOS)
# Core transcription
RealtimeSTT>=1.3.0

# System integration (Unix)
pyautogui>=0.9.50
keyboard>=0.13.5

# System tray support (optional)
pystray>=0.19.4
Pillow>=9.0.0

# HTTP client for AI enhancement
requests>=2.31.0

# Linux-specific packages
python3-xlib>=0.15; sys_platform == "linux"

# GPU acceleration (optional)
# torch>=2.0.0
# torchaudio>=2.0.0